ml prediction models/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
h11==0.16.0
idna==3.11
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
